Creates hash for all pieces' moves in piece
This means there won't need to be individual classes dedicated to the
different types of pieces, since the moves list is the only thing
differentiating them.

I noticed after creating the original Pieces class and changing King.rb
to inherit from it, that there wasn't much point in having the King
class, since Pieces (now Piece) could do everything for it and the other
pieces.