Corrects directions for queen, rook, and bishop
The northern and southern directions on their move lists were reversed.
This is because on a normal grid they would be right, but for the
program the "grid" is a 10x10 2D array, so for example you would need
to add to the y values to go down instead of subtract from them.